pyproject.toml
[tool.poetry]
name = "contractorium"
version = "0.1.0"
description = ""
authors = ["D0rs4n <41237606+D0rs4n@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^0.4.1"

[tool.poetry.group.dev.dependencies]
taskipy = "1.10.3"

[tool.poetry.group.lint.dependencies]
flake8 = "6.0.0"
flake8-annotations = "2.9.1"
flake8-bugbear = "^22.10.27"
flake8-docstrings = "1.6.0"
flake8-isort = "5.0.3"
flake8-string-format = "0.3.0"
flake8-tidy-imports = "4.8.0"
flake8-todo = "0.7"
pep8-naming = "0.13.2"
pre-commit = "2.20.0"

[tool.taskipy.tasks]
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
